LPS-144040 Extract auxiliary functions
